Fixed inconsistency with braces
Some parts of the tutorial went something(){, others went something() {. The former has no space between the statement and the opening brace; the latter does. Inconsistency of this kind is always bad, so I decided to go through the code and fix it up. I went with something() {, as this is much prettier (and more readable) than the alternative.

I've left the section of Allman-style code untouched, as I gathered that it's there to make a point regarding JavaScript's quirks when coding in Allman-style. Douglas Crockford argues that you should never use this style with JS (regardless of one's personal preference), because of these quirks. Perhaps in the future this section should be expanded to say so? In any case this time around I was more concerned with making the code consistent than anything else.

JH