re-order commands
I'm trying to perfect the order so that learning how to use tmux can be reasonably accomplished in one sitting. I feel like I may be leaving out some things that power users would be interested in, but for the sake of someone attempting to find out if they enjoy using tmux I believe it is in decent shape.