Your Name: Matt Subject Line: Addresses Comparisons in Javascript What Happened:
I believe that starting out with the double equals instead of the triple equals for strict comparison checking is incorrect. Because double equals uses type coercion, it is more of a feature the needs to be understood. Beginners looking at the language should look upon the stricter method as the proper one because it is less likely to give a surprising result.

I also tried to address the behaviour by adding an example to the double equals comparison.

Hope that the community is interested in pulling in these changes, they stem from teaching beginners javaScript but I am by no means the authority.