Remove misleading example of NA in Levels section
The levels section was using a level "NA" together with "female" and "male", which is misleading (given that NA was just introduced as a missing value). I updated it to what I believe is a better example of how NAs are interpreted.

Additionally, I edited a comment about Inf that wouldn't compile. How about inserting it at the tutorial altogether?