python3: changed to link to PYMOTW for python2
Python Module of the day http://pymotw.com/3 only has one module, so it might not be the best idea to link there. I changed it so that it links to 2 - but if that's not a good idea - maybe we should get rid of that link, until there actually is content there.