[hack/en] Fix link to other learnxinyminutes article
I had assumed a relative link to another markdown file in this project
would work, but it renders as:

    http://learnxinyminutes.com/docs/hack/php.html.markdown

instead of the desired:

    http://learnxinyminutes.com/docs/php/

I've replaced this relative link with an absolute link to where the target
page is located on the published site.

This commit also fixes a typo in markdown syntax at the end of the document.