Add an example of trap command  (#1826)
* Begin writing document for PCRE

Started writing learnxinyminutes document for PCRE to cover general purpose regular expressions. 
Added introduction and a couple of details.

* Change introductory example for regex

The old example was incorrect. It's replaced with a simple one.

* Add some more introductory text

* Add first example

* Added more example and a table for proper formatting

* Add few more examples

* Formatting

* Improve example

* Edit description of character classes

* Add a way to test regex

Add https://regex101.com/ web application to test the regex provided in example.

* Add example of trap command 

trap is a very important command to intercept a fatal signal, perform cleanup, and then exit gracefully. It needs an entry in this document.

Here a simple and most common example of using trap command i.e. cleanup upon receiving signal is added.

* Revert "Add example of trap command"

* Add an example of trap command

`trap` is a very important command to intercept a fatal signal, perform cleanup, and then exit gracefully. It needs an entry in this document.

Here a simple and most common example of using `trap` command i.e. cleanup upon receiving signal is added.