Added the Logical and bitwise operators section, fixes merge issue with  #1817 (#2292)
* Added the Logical and bitwise operators section

* Added a note for Short Circuit evaluation

Excerpt from https://en.wikipedia.org/wiki/Short-circuit_evaluation
C++ uses minimal evaluation, or McCarthy evaluation (after John McCarthy (computer scientist)) is the semantics of some Boolean operators in some programming languages in which the second argument is executed or evaluated only if the first argument does not suffice to determine the value of the expression: when the first argument of the AND function evaluates to false, the overall value must be false; and when the first argument of the OR function evaluates to true, the overall value must be true. In some programming languages (Lisp), the usual Boolean operators are short-circuit.