Change dates and times to actual occurances
The current dates were added by Adam Brenecki in the original commit. They
do not appear to have any real significance. This has been changed to the
following:

datetime: 2001-05-15T03:14:15.9Z

The date is when YAML started with a first pass specification. The time is
the first 6 digits of pi.

datetime_with_spaces: 2009-07-21 18:30:57.72 -1
The date is when the latest version of YAML was released (YAML 1.2) The time
is when a merge request for this commit will be created, followed by 5
digits of euler's constant

date: 2011-11-20
This is the latest news item on the website (http://yaml.org/). The news is
about JS-YAML, a JavaScript YAML parser.