Fixing tabs/spaces as per feedback received & some other minor changes
Fixing tabs/spaces as per feedback received & some other minor changes

```
Hi,
In case you still want feedback, when I open the tutorial in vanilla Chrome, the boxes with the examples wrap badly.  It looks as if a tab stop is too far to the right. It doesn't affect all the boxes equally -- the one about Constants is mostly OK.
In other feedback, that's exactly the level of tutorial I find helpful, so thanks for writing it!
```