Update rst.html.markdown
Revising the section on creating lists with RST, as the syntax for list/sublist creation is not 1-1 with Markdown. 

The current example is actually incorrect—any/all sublists must be separated from their parent list by blank lines (above and below). These sublists must also be properly indented in relation to the text of the above parent list.

References: 
[1]: http://docutils.sourceforge.net/docs/ref/rst/restructuredtext.html#bullet-lists
[2]: http://www.sphinx-doc.org/en/master/usage/restructuredtext/basics.html