[Raku/en] Fix runaway comment for syntax highlighter (#4215)
The pygments highlighter does not recognize `==begin` and `==end`
without a newline before `==end`. For this reason it will never end the
comment. This resulted in 95% of the Raku document being highlighted as
a comment.

We replace most `=begin` and `=end` with standard quotes `#`. In some
places I retain `=begin` and `=end` when they are indented (this works
fine in pygments).