Update set-theory.html.markdown
> A ∪ B = { x : x ∈ A ∪ x ∈ B }

I believe this is a circular definition. Defining an operator/operation using that same operation on both sides is pointless. Potentially, we could also define basic logical operations somewhere at the beginning (and, or, if-then, iff), cause they tend to be useful, or we could replace '∧' and others, with their standard names: and, or..